README.md
# Hermis
Hermis is a tool that gives users valuable information about the cost from the different airports to the city center.

## Inspiration
Have you ever found yourself with flights and hotel already booked using the amazing Skyscanner services but without taking into account the fees regarding public transport to the city center? That won't happen anymore with Hermis! Most of the time travelers choose an airport based on the price of the flight. Nevertheless, not considering the travelling costs can lead to an overall price higher than expected. With this tool we encourage to make a better decision by giving additional and useful information about the different public transport options and its correspondent price from the different airport options to the city center.

## How I built it
The tool is completely built with Python. It is based on Web Scraping and Text Processing using the libraries BeautifulSoup and Request. The data was obtained from the web: https://airmundo.com/ and text processing was developed by ourselves.

## Challenges I ran into

The main challenge was to learn how to use both the Web Scraping tools and the Skyscanner API. We were able to get a satisfactory result with the Web Scraping challenge but we did not have the same luck with the Skyscanner API. We were not able to use properly the JSON data so we could not implement the entire idea as it was planned in the beginning. 

In the social part, we did not have any results until 3 AM and motivation was not our main strength at that moment. 

## Accomplishments that I'm proud of

Surprisingly, we were able to motivate each other to start the whole project again from scratch at 3AM and start a non-stop-working environment that was fruitful. Understanding the basics of Web Scraping and HTML and being able to make a tool that accomplish our spectatives is also something we are very proud of.

## What I learned

We gathered a general knowledge about Web Scraping that would be extremely useful in the future. On the other hand, we discovered the potential that the API tools have and the amount of tools that companies make available.

## What's next for Hermis

Hermis has a lot of potential. Adding compatibility with Skyscanner is very interesting. At the moment, Skyscanner offers flights and hotels deals and this could be an interesting way to link them together. 
With this tool, the price ranking in Skyscanner could be more accurate including the expenses from the airport to the city center or the hotel.